+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 + Make `Terminal` generic.
 + Make `TermRead` more generic.
 + Implement `Eq` for `AmbigousEvent` and all sub structs.
++ Support control characters in `TermRead`.
 - Fix publicity of macro `codes::move_up`.
 - Macros in `codes` now evaluate their arguments only once. (also fixes
   affected codes in color macros)